requirements.txt
django==2.2.5
djangorestframework==3.10.3
django-short-url==1.1.4